Refactoring to allow import of config.Settings
